+++ brokerapi/brokers/cloudsql/cloudsql_suite_test.go
@@ -1,0 +1,13 @@
+package cloudsql_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestCloudsql(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "CloudSQL Suite")
+}